backend/requirements.txt
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
autopep8==1.7.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.1.1
fastjsonschema==2.16.2
Flask==2.2.2
Flask-Cors==3.0.10
google-api-core==2.10.2
google-auth==2.12.0
google-cloud-core==2.3.2
google-cloud-documentai==2.0.3
google-cloud-storage==2.5.0
google-cloud-translate==3.8.4
google-cloud-vision==3.1.4
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
graphviz==0.20.1
grpcio==1.49.1
grpcio-status==1.49.1
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.10
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter-server==1.21.0
jupyter_client==7.4.2
jupyterlab==3.4.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.1
notebook-shim==0.1.0
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pdf2image==1.16.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==4.21.7
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.13.0
PyMuPDF==1.20.2
pyparsing==3.0.9
PyPDF2==2.11.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
pyzmq==24.0.1
requests==2.28.1
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.1
terminado==0.16.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
